MAIL new SSA IL
The beginnings of Ssa2 are in place.  Here are what types in the new
IL look like.

	       Array of t
	     | Datatype of Tycon.t
	     | IntInf
	     | Object of {args: {elt: t, isMutable: bool} vector,
			  con: Con.t option}
	     | Real of RealSize.t
	     | Thread
	     | Vector of t
	     | Weak of t
	     | Word of WordSize.t

The main difference between Ssa2 types and Ssa types is the new
"Object" type, which is used for both constructed objects (with con =
SOME ...) and tuples (with con = NONE).  Each field in an object type
has an isMutable flag indicating whether or not it can be updated.
Ref types have also been dropped, since a ref is simply an object with
one mutable field.  The new Ssa2 types are different enough from the
old Ssa and XML types that I re-implemented the hash consing from the
ground up.

Here are what expressions in the new IL look like.

	       Const of Const.t
	     | Object of {args: Var.t vector,
			  con: Con.t option}
	     | PrimApp of {args: Var.t vector,
			   prim: Type.t Prim.t,
			   targs: Type.t vector}
	     | Profile of ProfileExp.t
	     | Select of {object: Var.t,
			  offset: int}
	     | Update of {object: Var.t,
			  offset: int,
			  value: Var.t}
	     | Var of Var.t

As with types, Object is used for both constructor and tuple
applications.  Similarly, select is used to select from both kinds of
objects.  Case transfers no longer implicitly select.  Finally, the
new Update expression is used to modify a field in both kinds of
objects.

I've put in the minimum amount of plumbing to get the new IL working.
So, there is the general analyzer and the type checker, and the passes
from Ssa to Ssa2 and from Ssa2 to Rssa, but nothing else.  The type
checker is like the old one, except the very simple notion of
subtyping that makes a constructed value a subtype of the datatype
that declares the constructor.  For now, there is no shrinker, and I
haven't even ported the ref flattening.  Both of those will happen
soon.

The pass from Ssa to Ssa2 is pretty simple.  It replaces Ref_assign
with Update, Ref_deref with Select, and Ref_ref with Object.  It
replaces ConApp and Tuple with Object.  It also adds selects to Case
targets, since those are no longer implicit.

I've only ported the -representation packed part of the backend to the
new IL.  I'm thinking it's not worth the effort to port the unpacked
stuff, which should be dropped since the packed stuff is better.

One interesting thing that this round of changes brought up was some
weaknesses in subtyping in the Machine IL.  There were some places
that unnecessarily required type equality instead of subtyping.