MAIL different layout for type error messages
I was thinking that type error messages would be easier to read if we
didn't put the file position and error description all on the same
line.  Instead, put the error description on its own line just after
the file position.  So, instead of

Error: z.sml 1.9: function not of arrow type
   function: [int]
   in: 1 2

We would get

Error: z.sml 1.9.
   Function not of arrow type.
   function: [int]
   in: 1 2

This isn't a great example since z.sml is such a small file name.
But, this approach has three visual advantages as I see it.  First,
the error description is always in the same place, instead of having
its start being dependent on the length of the file position.  Second,
the error description will be much less likely to go past 80 columns,
since it starts in column 3.  Third, since one usually uses automated
tools to put the cursor at the appropriate file position, one doesn't
even usually look at the file position anyways -- this makes it easier
to jump to what you want to read.

Anyways, this checkin implements the change.  Let me know what you
think.

Actually, with the extra space, I was thinking of writing out the
information on the file position, just to make it easier for beginners
(who might not use an automated tool), since I don't think it matters
for the rest of us.  Something like

Error in z.sml on line 1 column 9.
   Function not of arrow type.
   function: [int]
   in: 1 2

But I haven't checked that in yet.

Another format that I considered was the following, which redundandtly
displays "error", but is perhaps more uniform.

Error in z.sml on line 1 column 9.
   error: function not of arrow type
   function: [int]
   in: 1 2