Improvements to liveness computation and representation.  Used vectors instead of lists in places and used a Buffer.t instead of a list ref to accumulate variables.  This should save some space.  I was also more careful with not holding on to temporary information.
Added a simple shrinker for RSSA that inlines a goto to a block that
is only called once.  This cuts down on the number of blocks, which in
turn cuts down on the volume of liveness information.

Cleaned up limit check insertion and removed its assumption that in
its input, calls to functions (like IntInf_rem) that require bytes
free are in blocks by themselves.  We used to ensure this in the
SsaToRssa pass, but now the limit check pass does it itself.  It's a
lot cleaner, and is essential if we want to use the Rssa shrinker.