MAIL Added -inline-into-main {true|false}.
This flag controls whether the SSA inliner will inline functions into
the main function.  The idea was to use -inline-into-main false to
keep the size of main down.  Unfortunately, it doesn't work well,
because some stuff like determining the platform relies on some small
functions getting inlined into main.  This causes very basic constants
that one would expect to get folded away not to be.

I'll probably remove this flag soon since it's so useless.  This
checkin is mainly for documentation of a failed attempt.