MAIL fixed quadratic liveness for list constants
Fixed a performance problem that caused a quadratic amount of liveness
information for list constants.  The problem is that list elements are
evaluated left-to-right, but lists are constructed right-to-left.  So,
for [e1, e2, ..., en], the code looks like

 	x1 = e1
        x2 = e2
        ...
        xn = en
        x1 :: (x2 :: ... :: (xn :: []))

Note that at each ei, all of the previous xi are live.  Hence, if each
ei is a complex expression that cause the creation of a new basic
block, there will be a quadratic amount of liveness information.

The fix is to build the list in reverse as we go, and then reverse it
at the end.

	l0 = []
	l1 = e1 :: l0
	l2 = e2 :: l1
	...
	ln = en :: l(n-1)
	rev ln

With this approach, there is a constant number of live variables at
each block.

For now, I've hardwired in a constant, 20, so that lists larger than
that use the reverse trick, and smaller lists are evaluated directly.

This fix didn't help with the HOL performance problem, which is due to
the large number of blocks in main.