MAIL decreasing liveness information in large SSA functions
Implemented a pass ("zone") that runs at the end of the SSA simplifier
pipeline.  This pass breaks breaks large SSA functions into zones,
which are connected subgraphs of the dominator tree.  For each zone,
at the node that dominates the zone (the "zone root"), it places a
tuple collecting all of the live variables at that node.  It replaces
any variables used in that zone with offsets from the tuple.

The zones are chosen by walking down the dominator tree and
recursively "cutting" (choosing a zone root) at every n'th node.  At
some point, I may switch to an algorithm like the one I described
early that tries to make zones of equal size.

There are two flags that govern the use of this pass

	-max-function-size <n>
	-zone-cut-depth <n>

Zone splitting only works when the number of basic blocks in a
function is > n.  The n used to cut the dominator tree is set by
-zone-cut-depth.

For now, there is no attempt made to keep the zone roots outside of
loops.  For main, we rely on -inline-into-main false to ensure there
are no loops in main.  It should be easy enough to use loop forests to
fix this.  I've run loopForestSteensgard on Joe's main function
without any time problems.

There is currently no attempt to be safe-for-space.  That is, the
tuples are not restricted to containing only "small" values.


This hack significantly improves the liveness problems we've been
seeing in HOL.  For HOL, the RSSA program size is about 200M.  A
couple of days ago, we were at about 920M for the machine program
size, most of which is liveness info.  Switching to -inline-into-main
false cut that down to 750M.  With zone splitting, we're down to 250M,
which is acceptable, since there's always some blowup in going from
RSSA to machine due to things other than liveness.

The big remaining problem is compile time.  It takes 8.6 hours to
compile HOL.  Here's the breakdown.

      pre codegen finished in 456.25 + 260.84 (36% GC)
      x86 code gen finished in 24579.91 + 5485.26 (18% GC)
      Compile C and Assemble finished in 52.59 + 0.00 (0% GC)
      Link finished in 204.34 + 0.00 (0% GC)

The pre codegen only has one performance bug left for this program --
localRef is taking ~165 seconds (it should take ~5).  Hopefully this
is as easy as the quadratic problem with commonArg.  With that fix, we
should be at 10min for the pre codegen, which is acceptable.  I had
been worried that there would be a problem with the large C file for
main, which is about 8M.  There are also 74 assembly files, two of
which are quite large (30M and 50M), with most of the rest being
closer to the usual 0.5M.  The C compiler and assembler handle these
well.

Matthew, if you can look into why the x86-codegen is taking so long
that would be great.  Let me know if you need any other data.  Here
are the flags I compiled with.

	-basis 1997     \
	-keep g \
	-inline-into-main false \
	-verbose 3      \

One other piece of data.  It's not a property list problem.  Here are
the stats before and after the codegen.

before:
   numPeeks = 742821096
   maxLength = 12
   average position in property list = 0.081

after:
   numPeeks = 772932930
   maxLength = 12
   average position in property list = 0.115