MAIL Added val MLton.Exn.addExnMessager: (exn -> string option) -> unit
It is implemented in the General structure along with exnMessage.  The
implementation is as Matthew described.  Keep a (exn -> string option)
list ref of exception message extensions; addExnMessager adds a new
exception message to the list.  exnMessage iterates through the list
to see if there is an installed exception messager; if not, it returns
the exception name.

The messagers for IO and SysErr are installed immediately after each
of their exception declarations.  The messager for Fail had to be
delayed a bit after its definition, until concat is defined.