MAIL another improvement for list constants
If a list constant has zero or one expansive expressions, then build
the list right-to-left.

I thought of the following generalization that combines this hack with
the one I put in early to evaluate left-to-right, build the list in
reverse, and then reverse it in the end.

The idea is to evaluate all the expansive expressions left-to-right,
putting them in a list, and then to build the complete list
right-to-left, picking elements off the expansives list as we need
them.

So, for the following list, where ai is non-expansive and ei is
expansive

	[a0, e0, e1, a1, a2, e2, e3, e4, a3]

we would do

	r = []
	r = e0 :: r
	r = e1 :: r
	r = e2 :: r
	r = e3 :: r
	r = e4 :: r
	l = []
	l = a3 :: l
	l = hd r :: l
	r = tl r
	l = hd r :: l
	r = tl r
	l = hd r :: l
	r = tl r
	l = a2 :: l
	l = a1 :: l
	l = hd r :: l
	r = tl r
	l = hd r :: l
	r = tl r
	l = a0 :: l

Hopefully this makes the idea clear, as well as making it clear that
only a small constant number of variables are live.

Anyways, I haven't put this in, but if someone feels like it, the
right place is defunctorize.fun.