The vector type in SSA2 now has its element type as a product type, rather than a single type.  This will allow vector elements that are tuples to be flattened directly into the vector.
Eliminated the array type in SSA2, since it can be expressed as a
vector with mutable elements.

There's some code in ref flattening and in packed representations that
needs to be filled in to handle flattened vectors.  For now, that code
will raise an error.  It will need to be fixed as soon as I add a
vector flattening pass to SSA2.