Added Vector{Sub,Updates} expressions to Ssa2.
The representation pass now packs vector elements, and handles
Vector{Sub,Updates}.

Vector updates operate on a sequence of updates for two reasons.
First, the minor reason that if multiple updates occur to the same
vector element with pointers, then a single card mark sufficed for the
batch of updates.  Much more importantly, with packing, it may be much
more efficient to pack together several components, and update them
with a single store, than to perform each update separately (which
requires a load/modify/store per update).  The representation pass
doesn't take advantage of this yet, but it will.  Similar reasoning
applies to object updates -- I will modify that to take a sequence
soon.

The few remaining array primitives in SSA2 now take a type arg that
specifies the type of the entire array, not the element type.  This is
because the element type is a product type, which isn't a type in and
of itself, and hence can't be a type arg.