MAIL mlb integration
This commit brings an alternative "programming in the large" model to MLton.
It is essentially the model described at:
  http://www.mlton.org/pipermail/mlton/2004-March/015645.html
with some modifications to be consistent with Standard ML.

A good place to start is "Formal Specification of MLBs" in the User
Guide.  It is written in the style of the Definition (i.e., terse),
but gives a fairly detailed semantics for MLB behavior.  The only
missing piece there is annotations.  Steve, can you check that this
converts to HTML correctly? I don't have heava on any system I use.

Compiles of .sml and .cm files are transparently handled as implicit
MLB basdecs.  I've disabled the implicit compile of the basis file
when there are no file inputs and a def-use flag is set.  If you want
to compile the basis, do something like
  mlton -stop tc -show-basis basis.basis '$(SML_LIB)/basis/basis.mlb'

Please report bugs or suggestions.