MAIL reorganized Basis Library .mlbs
Reorganized the Basis Library implementation files so that
- $(MLTON_ROOT)/basis/basis{-2002}.mlb:
  Corresponds to strict adherence to the Basis Library specification
- $(MLTON_ROOT)/basis/basis-1997.mlb:
  Corresponds to strict adherence to the 1997 Basis Library specification;
  or, at least without extra signatures and structures below.
- $(MLTON_ROOT)/basis/mlton.mlb:
  Provides MLTON{_*} signatures and MLton structure.
- $(MLTON_ROOT)/basis/sml-nj.mlb:
  Provides SML_OF_NJ signature and SMLofNJ structure.
- $(MLTON_ROOT)/basis/unsafe.mlb:
  Provides UNSAFE signature and Unsafe structure.

For backwards compatibility, when compiling a .sml or .cm file, the
implicit MLBs BasDec looks like:
  local
    $(MLTON_ROOT)/basis/<<!Control.basisLibrary>>.mlb
    $(MLTON_ROOT)/basis/mlton.mlb
    $(MLTON_ROOT)/basis/sml-nj.mlb
    $(MLTON_ROOT)/basis/basis.mlb
  in
    z.sml
  end

There is one minor semantic change; compiling with -basis 1997 does
not provide the Primitive structure.  Hopefully no on was using it,
but it can be easily exported.

In light of the above, I've dropped compiler support for
-basis 2002-strict, and removed
$(MLTON_ROOT)/basis/basis-2002-strict.mlb.  I haven't nuked
/basis-library/libs/basis-2002-strict just yet, but there is no longer
a need for it.  On the other hand, I've kept support for -basis none
and $(MLTON_ROOT)/basis/basis-none.mlb, although I think that
$(MLTON_ROOT)/basis/pervasive.mlb is a better choice.