Found a gen-mlb.sml file that I wrote back in December, when we had first sketched out MLBs.  It was trivial to adapt to the current syntax.
However, it turns out that converting from a "pure" Portable Library
Description to an MLB leaves a little to be desired, because a
Portable Library Graph essentially leaves imported libraries as free
variables.  So, while it would be fairly straightforward to write an
pld2mlb utility (and gen-mlb.sml contains 90% of the logic for that),
it would necessarily have the usage:
  pld2mlb file.pld [import.mlb]*
to specify the imported libraries.  Conversely, a cm2pld utility (that
would essentially chain together CM.Graph.graph and
FormatPortable.output) would output a .pld with the imported libraries
left free.  It would be up to the user to inspect the .pld to discover
the correct ordering of the .cm files that were imported;
alternatively, cm2pld would need to output the imported .cm libraries
in some fashion.

In any case, this all seemed rather annoying, especially since
CM.Graph.graph returns the imported .cm libraries along with the
PortableGraph.graph.  So, cm2mlb simply grabs the .cm libraries,
converts them using the scheme below, and binds the resulting .mlb to
a basis identifier.

When cm2mlb encounters a .cm import, it attempts to convert it to a
.mlb file.  I've set things up so that cm2mlb automatically converts
$/basis.cm to $(MLTON_ROOT)/basis/basis.mlb (this changes the
semantics somewhat, because SML/NJ's $/basis.cm also provides SMLofNJ
and Unsafe structures; but if the code was using either of those
structures, you'll get a unbound identifier error when compiling with
mlton).  cm2mlb also checks ~/.mlton/cm2mlb-map for .cm to .mlb
mappings.  For example, I've put the following in mine:
  $/smlnj-lib.cm $(SMLNJ_ROOT)/smlnj-lib/Util/smlnj-lib.mlb
(where I've also added
  SMLNJ_ROOT /home/fluet/smlnj/smlnj-dist.cvs.HEAD/src
to my ~/.mlton/mlb-path-map).  You can also specify additional
cm2mlb-map files on the command line with -map file.

If the .cm file is not found in the cm2mlb map, then the file
extension is simply changed to .mlb.  This is often the "right thing"
for relative paths to sources.cm files.