Fixed a bug in the mlb cache (for both parsing and elaboration).
HashSet.lookupOrInsert should not be called with a thunk that may
mutate the HashSet.t, because the implementation captures the
HashSet.t state in a closure that is used to actually insert the new
element.  If the thunk modifies the HashSet.t, bad things can happen
(of particular concern, if the table is grown during the thunk, the
new element can be inserted into the wrong bucket (and with the wrong
tail), which means future lookups will not find the element (and
additional elements could be lost entirely)).

This was a problem with the mlb cache (which maps OS.FileSys.file_id's
to Ast.Basdec.t or Elaborate.Env.Basis.t), where upon encountering a
.mlb file for the first time, I recursively parsed or elaborated in
the insert thunk, which may encounter other .mlb files and mutate the
cache.

Instead, made the mlb cache map to Ast.Basdec.t Promise.t or
Elaborate.Env.Basis.t Promise.t, whereby the insert thunk immediately
returns with a Promise.t (making no recursive calls).  The Promise.t
is immediately forced, but after the HashSet.t is in a consistent state.