Some usability enhancements to cm2mlb.
- first OS.FileSys.chDir-s to the directory of the file to be converted;
  this ensures that relative paths are maintained

- made the imports of CM anchored paths more useful.  An anchored path
  is split into the anchor and the path; cm2mlb attempts to lookup the
  anchor in the cm2mlb-map; on success, concat the result path to the
  split path and change the extension to .mlb.  Otherwise, an
  unanchored path simply has it's extension changed to .mlb.

cm2mlb checks ~/.mlton/cm2mlb-map for .cm to .mlb
mappings.  For example, I've put the following in mine:
  smlnj-lib $(SMLNJ_ROOT)/smlnj-lib/Util
(where I've also added
  SMLNJ_ROOT /home/fluet/smlnj/smlnj-dist.cvs.HEAD/src
to my ~/.mlton/mlb-path-map).  You can also specify additional
cm2mlb-map files on the command line with -map file.

Note that in the above case, the anchor is smlnj-lib, which is often
used in .cm files as $/smlnj-lib, which is just an abbreviation for
$smlnj-lib/smlnj-lib.cm.  So, it is an anchored path, is split into
(smlnj-lib, smlnj-lib.cm), and the corresponding .mlb that appears in
the output is $(SMLNJ_ROOT)/smlnj-lib/Util/smlnj-lib.mlb.