Fixed a long-standing bug in the translation of the GC_collect primitive into a C call to GC_gc.  It had been incorrectly translated with maySwitchThreads = false.  It is now translated with maySwitchThreads = handlesSignals.  Fortunately, most of the calls to GC_gc are inserted by the limit check pass, which had the correct maySwitchThreads.
This bug showed up when I compiled a program that called
MLton.GC.collect with -gc-check first.  This caused the GC_gc call to
switch threads but the code to continue on as if it were in the
original thread, leading to a segfault.

I guess this could be checked for in the Rssa and Machine type
checkers.