MAIL most MinGW regressions now pass
I'll go over the failures tomorrow with a fine-tooth comb and figure
out which are OK and which need fixing.

Cleaned up the runtime Makefile's handling of the C files and
eliminated the duplication in OBJS and DEBUG_OBJS.  Added the ability
to compile the runtime as one large C file with COMPILE_FAST=yes.  The
cross regression script uses this to speed up testing.

Cleaned up how unimplemented system calls are stubbed out in the basis
library code.  Now there is a single file, stub-mingw.sml, that
wraps stubs around all the unimplemented calls.

Improved the regression script so that it can cross compile the mingw
tests.