A broken removeUnused for SSA2.
The problem is that constructor types can exist independently of their
datatypes.  So, replacing unused constructors with dummy constructors
requires changing the type of the variables to which they are bound.
Previously, removeUnused was trying to avoid needing to rewrite types,
but I think we need to do so in order to flush out unused constructors.