MAIL added MLton.share
	val MLton.share: 'a -> unit

MLton.share x maximizes sharing in the heap for the object graph
reachable from x.  It is implemented by the GC_share runtime function.
It works by using hash consing, inserted at two places in the
mark-compact depth-first traversal code.

1. As the traversal of each object is finished, its contents are
hashed and looked up in the hash table.  If an equivalent object is
already in the table, then that object is returned.  Otherwise, the
object just finished is inserted.

2. When an already marked pointer is encountered, the contents of the
object pointed to are hashed and looked up in the table.  Since the
object is marked, we know some equivalent object will be there (it may
be the object itself), so we replace the pointer with that object.

In order for the runtime system to respect SML's notion of
equivalence, it must have information about which objects require
identity to be preserved (i.e. ref cells and arrays).  So, I added
an additional field to GC_objectType, a boolean "hasIdentity", and
propagated hasIdentity information through the backend into the
codegens.

Right now, GC_share uses temporary malloc'd space outside the SML heap
for its hash table.  I plan to add code to use free space at the end
of the heap if it is available.  However, once that space is used up,
we either have to allocate additional memory or let the quality of the
hash-consing degrade.  I'm not sure which is best.

The rehashing of already-marked objects in (2) annoys me a bit.  The
only way I see to alleviate that is to use an additional hash table,
keyed on pointers.  I'm not sure if this is worth it, either in code
complexity or time savings.

Right now, only fixed-size objects are hashed.  It may be worthwhile
to handle vectors as well.

Next week, I plan to add

	val MLton.shareAll: unit -> unit

which will maximize sharing over the entire heap.  I've already put
the code in the basis library and a stub in gc.c.  It shouldn't take
too much to fill it out using ideas similar to the above -- i.e. do a
mark-compact gc over the whole heap, with a little hash consing as
objects are finished.  In this case, I may be able to avoid the extra
hashing of (2) and an extra table by using a trick similar to forward
pointers.  The reason this works with shareAll and not share is that
with shareAll we can trash objects that are not inserted into the hash
table, since we know they will be unreachable.  With share, there may
be external pointers into the object graph, so we can't trash
anything.

Henry, since you've implemented all this before in Mathematica, I'd
appreciate any comments you have.