The peephole optimization that removes a self-move may invalidate the recorded liveness of indirect operands.  For example, consider
 live-in: RP(2), RP(3), RP(7)
 movl OP(RP(2), 0), OP(RP(2), 0)
 live-out: RP(4), RP(7)

When the redundant instruction is eliminated, then RP(2) is no longer
in the live-in set.  To restore the consistency of the liveness
information, we recompute the liveness of the block.  Since self-moves
of temporaries are common (and the motivation for the optimization),
we avoid recomputing the liveness when the eliminated memory location
does not utilize other memory locations that have their liveness
tracked.