MAIL C types now distinguish between signed and unsigned words
This distinction is necessary because on some platforms the calling
convention for signeds and unsigneds is different.  This can happen
when a small word (e.g. 8 bit) is represented in a larger word
(e.g. 32 bit), in which case the signed version will be passed sign
extended and the unsigned version will be passed zero extended.

One unfortunate side effect of this change is that the type
information for a C function (CFunction.t) is now duplicated, because
we need the CType information to generate the prototype, while we need
different type information for type checking the ILs.  It's not clear
if this duplication is worth the gain in type checking, especially
since the C prototype can lie anyways.  But, I've left it in for now.
Of course, another fix would be to make the IL type information
sufficiently precise that it subsumes the CType information.  But
that's a bit more pervasive of a change than I'm willing to stomach
right now, plus I think it makes arbitrary distinctions between signed
and unsigneds.

A quick check reveals that generated C code now has correct prototypes
for the Word_{quot,rem} functions.

  Int8 WordS8_quot (Int8 x1, Int8 x0);
  Int8 WordS8_rem (Int8 x1, Int8 x0);
  Int32 WordS32_quot (Int32 x1, Int32 x0);
  Int32 WordS32_rem (Int32 x1, Int32 x0);

Filip, let us know if the fixed-integer regression now works without
needing any explicit sign extensions, or if this checkin has caused
any other problems.