MAIL fixed IntInf.size bug that leads to a seg fault
size now returns 1 on small ints, as it used to.

MLton.IntInf.size is no longer exported.

Added an assert in the end of setFrontier() in IntInf.c to ensure that
we never set the frontier beyond the allotted bytes.  I think that
check is sufficiently strong to catch the bug that we don't need to
take further measures like telling gmp not to realloc.