MAIL support for constants defined on the command-line
Added command-line switch "-const '<name> <value>'", which is used to
set the value of constants defined in SML by _command_line_const.  The
syntax is

	_command_line_const "<name>": <ty> = <const>;

This is now used for constants

	Exn.keepHistory, MLton.safe, MLton.size, TextIO.bufSize

The -const switch replaces the following, which are now expert and
deprecated:

	-detect-overflow
	-exn-history
	-safe
	-text-io-buf-size

One annoyance is that the value of exnHistory is used inside the
compiler itself, in the implement exceptions pass.  Perhaps this could
be removed, but it's not obvious.  In any case, there is a little
hardwired code that sets the value of Control.exnHistory based on the
value of "Exn.keepHistory".  I don't think this is any worse than
before, when we had hardwired code to set the build constant named
"Exn_history" based on the value of Control.exnHistory.  The other
three constants were unused in the compiler, so their control was
removed.