MAIL refFlatten fix and improvement
Fixed a bug in refFlatten.  The bug is now tested for by the
regression ref-flatten.3.sml.  The bug was that refFlatten mistakenly
flattened a ref cell allocated prior to a loop into an object
allocation inside the loop, thus losing sharing of the ref.  The fix
was to require the object allocation to be in the same block as the
ref allocation.  This is pretty draconian, and it would be nice to
generalize it some day to allow flattening as long as the ref
allocation and object allocation "line up one-to-one" in the same
loop-free chunk of code.

I also put in Matthew's suggested improvement, which allows updates to
the ref cell provided that they occur in the same basic block in which
it is allocated (and hence it is safe-for-space because the containing
object is still alive).  It would be nice to relax this to allow
updates as long as it can be proved that the container is live.

Added several refFlatten regression tests.