MAIL spaces in mlb imported file names
Extended the MLB lexer to parse SML style strings, which are passed as
FILE tokens to the MLB parser.  This allows arbitrary characters in
file names.  Note, however, that  mlton -stop f  will print file names
without quoting, so spaces in file names break using  mlton -stop f
in Makefiles.

Note that the MLB lexer has a fairly limited regexp for filenames.
Filenames must be composed of alpha-num characters or "-" or "_".  So,
you need to "quote" for anything more advanced.

I'm not 100% happy with this situation, but I wasn't very successful
in determining the Makefile or bash filename regexp.  Also,
considering that this is SML, I'm tempted to use SML style strings,
which makes it easy for a user to understand what they are writing
down, although it makes writing Win32 filenames a bit verbose.