MAIL Loading dynamic libraries works on Darwin
It works because Darwin has a "dlcompat" compatibility layer built on
top of their native dyld stuff.  For the iimport.sml test, I needed to
change the filename of the library.  Also, Darwin doesn't like it if
you call dlclose on a dynamic library -- it reports the error message
"dynamic libraries cannot be closed".  So, I disabled dlclose on
Darwin.

I wonder if we should make the iimport.sml DynLink stuff available in
the MLton structure.