MAIL added MLton.CallStack
Its main function,

	val MLton.CallStack.current: unit -> t

returns a representation of the current call stack.  There is also

	val toStrings: t -> string list

that takes a call stack and returns the corresponding list of
functions (name and source position) on the stack.

MLton.CallStack.current is implemented by calling a runtime function
that walks the entire stack, and stores the frameIndex for each frame.
The internal representation of the call stack is an int array (of the
frame indices).

To convert the call stack to a string list, each frame index is
converted into its corresponding sequence of source function indices,
and then each source function index is converted to its corresponding
string.  This relies on having at runtime the information hitherto
only used for profiling.  This information is created by inserting
profiling annotations into the source program just as with other forms
of profiling.  The only difference is that the profile pass on the
Rssa doesn't insert any profile labels (as it would with -profile
time) or calls to increment functions (as it would with -profile
{alloc,count}).  Instead, it merely creates the frame data to be used
at run time.

Added a new flag, -profile call, which governs whether call stack
information is available at runtime.  This is a bit of a misnomer,
since it doesn't do profiling, but it seemed like the best fit.  I am
open to improvements.

Used MLton.CallStack to improve MLton.Exn.history.  Exn.keepHistory
true now implies -profile call.  The first time an exception is
raised, we call MLton.CallStack.current and store the result in the
raised exception.  Exn.history simply extracts the call stack and
converts it to a string list.  The output of this is a vast
improvement over what was there before.  However, there is a cost --
raising an exception now takes time proportional to the size of the
call stack instead of constant time.  Of course, this is only with
Exn.keepHistory true, which will probably only be used for debugging.
I can imagine trickier implementations that have better performance,
only walking the stack up to the handler, and walking the rest if the
exception is later re-raised.  A user could also roll their own call
stack stuff (say, for Assert.assert), where the program is compiled
with -profile call, but with Exn.keepHistory false, and only grabs the
call stack when there is an exception that is really an error, as
opposed to an exception used for some other control flow.  Another
approach would be for us to provide special raise functions that {do,
do not} grab the history.

Anyways, I'm not sure which of this performance-improving approaches
is the right way to go, and since I'm not sure if they're needed, I'll
hold off for now.

Because of the new approach to MLton.Exn.history, both the CoreML and
Xml ILs were simplified to eliminate the "filePos" component of raise
expressions.  Those had been used by the old mechanism to propagate
the file position information.  The new approach is much better.

There was some cleanup of implementExceptions, making it a bit more
general.  It used to incorrectly syntactically move the argument to
setExtendExtra.