Library improvements.
Exported structure Position.

Changed the type few String functions.  They were all along the lines
of replacing String.isSuffix {string, suffix} with String.hasSuffix
(string, {suffix}).  It's more consistent with our convention to keep
the main argument unnamed and out of the record.  The record for the
other argument suffices to distinguish which string is which.

Improved structure Random, in particular, the function for choosing a
random number in a certain range.  There is no longer any
precomputation needed.

Created http.mlb to export Http structure.

Renamed Uri to Url.

Improved structure Http.

Reordered the results of Queue.deque to make the queue first, which is
more in keeping with our convention.