MAIL integrated Wesley Terpstra's patch for MLton.Process.create.
Reorganized how the various platforms implement mmapAnon, either via
an underlying mmap or VirtualAlloc.  This was to clean up some #if's
from gc.c and, more importantly, to make it easier to add the runtime
switch use-mmap.  This switch, whhich is only meaningful on Cygwin,
controls whether a Cygwin executable uses mmap or VirtualAlloc for
memory allocation.  If an executable uses mmap, then fork is enabled
and Process.create will use the usual Unixish stuff.  On the other
hand, if an executable uses VirtualAlloc, then fork is disabled and
Process.create will use the underlying windows CreateProcess.