Preliminary support for importing the address of symbols with the syntax:
  _import # "label" : MLton.Pointer.t;

Under this scheme, the following are almost equivalent:

  val f = _import "foo" : int -> int ;

and

  val f_addr = _import # "foo" : MLton.Pointer.t ;
  val f_schema = _import * : MLton.Pointer.t -> int -> int ;
  val f = f_schema f_addr

Likewise, the following are equivalent:

  val f = _import "foo" : int ;

and

  val f_addr = _import # "foo" : MLton.Pointer.t ;
  val f = MLton.Pointer.getInt32(f_addr, 0)

The elaboration of FFI related constructs could use a bit of an
overhaul.  I don't see any reason not to import a non-functional type
as it's address followed by a get of the appropriate type (the x86
codegen is doing exactly this, and I imagine that the C codegen is as
well).  I still maintain that it would be better to import a (direct)
function _not_ as it's address followed by an application of the
appropriate indirect function call schema, because a direct function
call is better for processor branch prediction, etc.

I didn't put much thought into syntax; I choose "#" because it was a
symbol known by the lexer.