MAIL bugfix: Bug with "polymorphic" exns
Fixed a front end bug that incorrectly disallowed expansive valbinds
that bind type variables (implicitly or explicitly, it doesn't matter)
when the type variable doesn't occur in the type of the value being
bound.  For example, the following were incorrectly rejected.

  val x = let exception E of 'a in () end

  val 'a x = let exception E of 'a in () end

The elaborator had simply checked if the right-hand side was
expansive, and if so, and there were bound type variables, then it
reported an error.  Now, it only reports an error if bound type
variables occur in the type of the right hand side.

I left the output of the elaborator unchanged.  The conversion from
CoreML to Xml handles rearranging the type variable binding.  For
example, it turns

  val 'a x = let exception E of 'a in () end

into

  val x = let
             val 'a f = fn () => let exception E of 'a in () end
          in
             f[unit] ()
          end

This effectively implements Matthew's suggestion to instantiate the
unused type variables with unit.

In making this change, I noticed and fixed a bug in unification of
flex records.  I couldn't think of a program that would tickle it
though.

There are still some programs that are incorrectly rejected.  For
example, although

  val 'a f = let exception E of 'a in E end

should be rejected and is rejected,

  val 'a _ = let exception E of 'a in E end

should be accepted, but is rejected.  The reason it should be accepted
is that even though the bound type variable ('a) appears in the type
of the right-hand side ('a -> exn), it doesn't appear in the type of
any of the variables bound by the left-hand side.  I think a simple
fix will get this too, but I wanted to go ahead and commit, since the
current fix passes all regressions, including a new one that tests
this stuff: expansive-valbind.sml.