MAIL bugfix: grammar for programs
Fixed a bug in the grammar for programs that prevented sequences of
expressions (separated by semicolons) after a topdec had already been
processed.  For example, the following program was incorrectly
rejected.

   val x = 0;
   ignore x;
   ignore x;

Thanks to Andreas Rossberg for reporting the bug.

The fix lead to much cleaner productions for programs than what we had
before.  Here are (essentially) the new productions.

  program ::= exp ; program
            | topdecs
  topdecs ::=
            | topdec topdecs
            | ; program

(exp and topdec had the same productions as before)