Fixed a bug that caused the internal bug message
	close useBeforeDef

to be reported instead of reporting the type error:

	Type escapes the scope of its definition

This bug was recently introduced in the change in the implementation
of type environment closure, which needed to tick the time clock, but
didn't have a proper "useBeforeDef" reporter to install.  I tweaked
elaborate-core to pass in the reporter from the appropriate
declaration.