MAIL bytecode codgen working again
Type synonym replacement in the front end tickled a bug in the
bytecode codegen which could result in the following error:

missing opcode: Int32_storeStackOffset

The bytecode abstract machine treates all words as unsigned, but
foreign C-calls must respect the signed/unsigned distinction to
properly handle calling conventions.  Hence, using the C-prototype to
derive bytecode operations may introduce signed data, for which there
are no corresponding opcodes or registers.  We mediate in the
generated C-call dispatch with explicit casts to and from signed words
when the C-prototype demands.